Reorganize database files into internal/database package
