+++ backend/internal/database/init_db.sql
